Create ARP spoofer
This tool is used by the attacker to become the MITM by fooling the target's machine into thinking that the attacker's machine is the router, then fooling the router into thinking the attacker's machine is the Victim. That way all the traffic going between the victim and the router is intercepted by the attacker.